refactor: 게시물 수정 dialog
