+++ src/widgets/post/edit-dialog/index.tsx
@@ -1,14 +1,28 @@
 import { useSelectedPost } from "@/entity/post/model/selectedPost"
 import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea } from "@/shared/ui"
 import { useDialogStore } from "@/features/dialog/model/store"
-import { useEditPost } from "./model/useEditPost"
+import { useMutation } from "@tanstack/react-query"
+import { editPost } from "@/entity/post/api/editPost"
 
 export const EDIT_POST_DIALOG = "EDIT_POST_DIALOG"
 
 export const EditPostDialog = () => {
   const { selectedPost, setSelectedPost } = useSelectedPost()
-  const { isDialogOpen, setDialogOpen } = useDialogStore()
-  const { mutation } = useEditPost()
+  const { isDialogOpen, setDialogOpen, closeDialog } = useDialogStore()
+
+  const mutation = useMutation({
+    mutationFn: (postId: number) => {
+      return editPost(postId, selectedPost)
+    },
+    onSuccess: (data) => {
+      closeDialog(EDIT_POST_DIALOG)
+      // 기존 게시물 목록 수정
+      console.log(data)
+    },
+    onError: (error) => {
+      console.error("게시물 수정 오류:", error)
+    },
+  })
 
   if (!selectedPost) return null
